requirements.txt
numpy
tensorflow-gpu
tqdm
